+++ CHANGELOG.md
@@ -6,6 +6,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic
 Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [6.10.0] - 2018-10-25
+
+#### Added
+
+- Added Content-MD5 checks for s3 file uploads
+
 ## [6.9.0] - 2018-10-12
 
 #### Added